created vrms-event.js and moved sortByDate and timeFomat funcs to that file and call appendMeetingTimes from vrms-events now
