+++ assets/js/project.js
@@ -1,12 +1,13 @@
 ---
 ---
-
 /* 
-  Fetch the correct project
+Fetch the correct project
 */
 {% assign projects = site.data.external.github-data %}
 // Escapes JSON for injections. See: #2134. If this is no longer the case, perform necessary edits, and remove this comment.
 let projects = JSON.parse(decodeURIComponent("{{ projects | jsonify | uri_escape }}"));
+
+import { vrmsDataFetch, timeFormat } from './utility/vrms-events.js';
 
 /*
   Passing script attributes from html script tag to JS file
@@ -136,58 +137,8 @@ let meetingsHeader = document.querySelector('.meetingsHeader');
 let projectTitle = scriptTag.getAttribute("projectTitle");
 let meetingsFound = [];
 
-// Grab the meeting time data from the vrms_data.json file
-{% assign vrmsData = site.data.external.vrms_data %}
-// Escapes JSON for injections. See: #2134. If this is no longer the case, perform necessary edits, and remove this comment.
-const vrmsData = JSON.parse(decodeURIComponent("{{ vrmsData | jsonify | uri_escape }}"));
-
-// me 
-console.log("vrmsData", vrmsData)
-
-// Helper function to sort VRMS data by day of the week from "date" key and meeting time from "startTime" key
-function sortByDate(scheduleData) {
-    console.log("scheduleData", scheduleData)
-    const map = {
-        'Mon': 1,
-        'Tue': 2,
-        'Wed': 3,
-        'Thu': 4,
-        'Fri': 5,
-        'Sat': 6,
-        'Sun': 7
-     };
-
-     scheduleData.sort(function(a, b) {
-        const day1 = new Date(a.date).toString().substring(0, 3);
-        const day2 = new Date(b.date).toString().substring(0, 3);
-
-        return map[day1] - map[day2];
-     });
-
-     scheduleData.sort(function(a, b) {
-        const day1 = new Date(a.date).toString().substring(0, 3);
-        const day2 = new Date(b.date).toString().substring(0, 3);
-        const time1 = new Date(a.startTime).toString().substring(16, 21);
-        const time2 = new Date(b.startTime).toString().substring(16, 21);
-
-        if (day1 === day2) {
-            if (time1 > time2) {
-                return 1;
-            } else {
-                return -1;
-            }
-        } else {
-            return 1;
-        }
-     });
-}
-
 // Loops through the VRMS data and inserts each meeting time into the HTML of the correct project page
 function appendMeetingTimes(scheduleData) {
-    
-    sortByDate(scheduleData);
-    //me 
-    console.log("scheduleData ====>", scheduleData)
 
     for (const event of scheduleData) {
         try {
@@ -196,6 +147,7 @@ function appendMeetingTimes(scheduleData) {
             const projectName = event.project.name;
             const name = event.name;
             const day = new Date(event.date).toString().substring(0,3);
+
             // only append the meeting times to the correct project page
             if (projectTitle.toLowerCase() === projectName.toLowerCase()) {
                 meetingsList.insertAdjacentHTML("beforeend", `<li class="meetingTime">${day} ${startTime} - ${endTime} <br>${name}</li>`);
@@ -208,35 +160,10 @@ function appendMeetingTimes(scheduleData) {
     } 
 }
 
-appendMeetingTimes(vrmsData);
-
+vrmsDataFetch("project", appendMeetingTimes)
 
 if (meetingsFound.length >= 1) {
     meetingsHeader.insertAdjacentHTML("beforeend", '<strong class="meetings">Meetings</strong>');
     document.querySelector('#userTimeZone').insertAdjacentHTML('afterbegin', timeZoneText());
-
-}
-
-// Formats time to be readable
-function timeFormat(time) {
-    let hours = time.getHours();
-    let minutes = time.getMinutes();
-
-    if (minutes == 0) {
-        minutes = minutes + "0";
-    }
-
-    if (hours < 12) {
-        return `${hours}:${minutes} am`;
-    }
-    else if (hours > 12){
-        hours = hours - 12;
-        return `${hours}:${minutes} pm`;
-    }
-    else if (hours = 12){
-        return `${hours}:${minutes} pm`;
-    } else {
-        return `${hours}:${minutes} am`;
-    }
 
 }
